fix(runtime): retain original reconciliation file layout
Keep the abandoned-start fix and regression tests in existing files. Restore the original launcher helpers and remove module extraction and documentation changes so the PR shows only the runtime correction.
