+++ sdk/rust/lib/backend/local/sandbox/mod.rs
@@ -6,7 +6,6 @@
 //! methods; [`LocalBackend::create_sandbox`] is its entry point.
 
 mod create;
-mod reconciliation;
 
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::path::Path;
@@ -15,6 +14,7 @@ use std::time::Duration;
 
 use base64::{Engine as _, engine::general_purpose::URL_SAFE_NO_PAD};
 use futures::{StreamExt, future::BoxFuture, stream};
+use microsandbox_db::pool::DbPools;
 use microsandbox_db::{DbReadConnection, DbWriteConnection};
 use microsandbox_image::{Digest, GlobalCache};
 use microsandbox_protocol::message::MessageType;
@@ -44,6 +44,7 @@ use crate::sandbox::{
     SandboxStatus, load_sandbox_record, validate_env, validate_hostname, validate_labels,
     validate_volume_mounts,
 };
+use create::{SandboxTransitionGuard, sandbox_runtime_endpoint_is_live};
 
 //--------------------------------------------------------------------------------------------------
 // Constants
@@ -538,6 +539,167 @@ impl LocalBackend {
 //--------------------------------------------------------------------------------------------------
 
 impl LocalBackend {
+    /// Reconcile a Starting/Running/Draining row against the owning process's
+    /// liveness, marking it terminal when the runtime is gone.
+    async fn reconcile_sandbox_runtime_state(
+        &self,
+        pools: &DbPools,
+        sandbox: sandbox_entity::Model,
+    ) -> MicrosandboxResult<sandbox_entity::Model> {
+        let run_dir = self.config().run_dir();
+        let sandboxes_dir = self.config().sandboxes_dir();
+        Self::reconcile_sandbox_runtime_state_with_paths(
+            pools,
+            sandbox,
+            Some((&run_dir, &sandboxes_dir)),
+        )
+        .await
+    }
+
+    /// Reconcile runtime state with optional exact socket roots.
+    async fn reconcile_sandbox_runtime_state_with_paths(
+        pools: &DbPools,
+        sandbox: sandbox_entity::Model,
+        socket_roots: Option<(&Path, &Path)>,
+    ) -> MicrosandboxResult<sandbox_entity::Model> {
+        Self::reconcile_sandbox_runtime_state_with_transition(pools, sandbox, socket_roots, None)
+            .await
+    }
+
+    /// Reuse a start caller's transition ownership instead of trying to lock itself.
+    async fn reconcile_sandbox_runtime_state_with_transition(
+        pools: &DbPools,
+        sandbox: sandbox_entity::Model,
+        socket_roots: Option<(&Path, &Path)>,
+        held_transition: Option<&SandboxTransitionGuard>,
+    ) -> MicrosandboxResult<sandbox_entity::Model> {
+        if !matches!(
+            sandbox.status,
+            SandboxStatus::Starting | SandboxStatus::Running | SandboxStatus::Draining
+        ) {
+            return Ok(sandbox);
+        }
+
+        let run = Self::load_active_run(pools.read(), sandbox.id).await?;
+        if run
+            .as_ref()
+            .and_then(|run| run.pid)
+            .is_some_and(Self::pid_is_alive)
+        {
+            return Ok(sandbox);
+        }
+
+        // A launcher owns Starting before the child records its PID. Prove that
+        // handoff abandoned with the same lock used by existing launchers.
+        let transition_guard = if sandbox.status == SandboxStatus::Starting
+            && held_transition.is_none()
+            && let Some((run_dir, _)) = socket_roots
+        {
+            let Some(guard) = Self::try_acquire_sandbox_transition_guard(run_dir, &sandbox.name)?
+            else {
+                return Ok(sandbox);
+            };
+            Some(guard)
+        } else {
+            None
+        };
+        let owns_transition = held_transition.is_some() || transition_guard.is_some();
+
+        // A dead-PID snapshot is not sufficient: another process may already
+        // have reconciled and restarted this name. Serialize on the runtime
+        // ownership lock, then re-read the exact row/run before unlinking.
+        let _guard = if let Some((run_dir, _)) = socket_roots {
+            let Some(guard) =
+                microsandbox_runtime::ipc::try_acquire_lifecycle_guard(run_dir, &sandbox.name)?
+            else {
+                return Ok(sandbox);
+            };
+            Some(guard)
+        } else {
+            None
+        };
+        let Some(sandbox) = sandbox_entity::Entity::find_by_id(sandbox.id)
+            .one(pools.read())
+            .await?
+        else {
+            return Err(crate::MicrosandboxError::SandboxNotFound(sandbox.name));
+        };
+        if !matches!(
+            sandbox.status,
+            SandboxStatus::Starting | SandboxStatus::Running | SandboxStatus::Draining
+        ) {
+            return Ok(sandbox);
+        }
+        let run = Self::load_active_run(pools.read(), sandbox.id).await?;
+
+        // Only an unowned start can be declared abandoned without a run row.
+        // Older runtimes may expose an endpoint without holding lifecycle locks.
+        let Some(run) = run else {
+            let abandoned_start = sandbox.status == SandboxStatus::Starting && owns_transition;
+            if abandoned_start
+                && let Some((run_dir, sandboxes_dir)) = socket_roots
+                && sandbox_runtime_endpoint_is_live(
+                    run_dir,
+                    &sandboxes_dir.join(&sandbox.name),
+                    &sandbox.name,
+                )?
+            {
+                return Ok(sandbox);
+            }
+            if sandbox.status == SandboxStatus::Draining || abandoned_start {
+                if let Some((run_dir, sandboxes_dir)) = socket_roots {
+                    crate::runtime::remove_sandbox_socket_artifacts_at(
+                        run_dir,
+                        sandboxes_dir,
+                        &sandbox.name,
+                    )?;
+                }
+                let (terminal_status, reason) = Self::stale_runtime_terminal_state(sandbox.status);
+                Self::mark_sandbox_runtime_stale(
+                    pools.write(),
+                    sandbox.id,
+                    None,
+                    terminal_status,
+                    reason,
+                )
+                .await?;
+
+                return sandbox_entity::Entity::find_by_id(sandbox.id)
+                    .one(pools.read())
+                    .await?
+                    .ok_or_else(|| crate::MicrosandboxError::SandboxNotFound(sandbox.name));
+            }
+
+            return Ok(sandbox);
+        };
+
+        if run.pid.is_some_and(Self::pid_is_alive) {
+            return Ok(sandbox);
+        }
+
+        if let Some((run_dir, sandboxes_dir)) = socket_roots {
+            crate::runtime::remove_sandbox_socket_artifacts_at(
+                run_dir,
+                sandboxes_dir,
+                &sandbox.name,
+            )?;
+        }
+        let (terminal_status, reason) = Self::stale_runtime_terminal_state(sandbox.status);
+        Self::mark_sandbox_runtime_stale(
+            pools.write(),
+            sandbox.id,
+            Some(run.id),
+            terminal_status,
+            reason,
+        )
+        .await?;
+
+        sandbox_entity::Entity::find_by_id(sandbox.id)
+            .one(pools.read())
+            .await?
+            .ok_or_else(|| crate::MicrosandboxError::SandboxNotFound(sandbox.name))
+    }
+
     /// Load the most recent active run record for a sandbox, if any.
     pub(crate) async fn load_active_run(
         db: &DbReadConnection,
